mm: vm_sequence is protected by semaphore
The vm_sequence is protected by the mmap_sem (and potentially by the
i_mmap_sem for file mapping), but the seqcount API wants the writers to be
protected against preemption which is not need because of the semaphore.

There is no seqcount API for semaphore, and that will not match our case
because some operation are not protected by the mmap_lock but using the
i_mmap_sem managed by the filesystem. This means the lock attached to the
seqcount would have to be change in the case of a file mapping. That's not
a big issue but useless today as the file mapping are not handled using the
speculative page fault handler.

Using raw API to prevent lockdep to complain until seqcount API is dealing
with semaphores and the attached semaphore is managed correctly for file
mapping.

Signed-off-by: Laurent Dufour <ldufour@linux.ibm.com>