x86/mm: try vma lock-based page fault handling first
Attempt vma lock-based page fault handling first, and fall back to the
existing mmap_lock-based handling if that fails.

The vma is looked up and locked under protection of the rcu read lock.
Because vma tree lookup under rcu does not guarantee that we retrieved
the correct vma, we perform additional check after vma is locked.

Signed-off-by: Suren Baghdasaryan <surenb@google.com>